move functions to conreq.utils.multiprocessing
